requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorlog==6.9.0
confection==0.1.5
cryptography==45.0.2
cymem==2.0.11
Deprecated==1.2.18
Events==0.5
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.2.6
opensearch-py==2.8.0
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pikepdf==9.7.0
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pymupdf4llm==0.0.24
pypdf==5.5.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
unicodedata2==16.0.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
